Relocate object store requests
